+++ CHANGELOG.md
@@ -12,7 +12,10 @@ Note: Can be used using nvuillam/mega-linter@insiders in your mega-linter.yml fi
   - Add examples of success and failed linter logs in documentation
   - Remove `/tmp/lint` and `/github/workspace` from log files
 
-- Add list of supported IDE in each linter documentation
+- Documentation
+  - Add list of supported IDE in each linter documentation
+  - Generate GitHub card on linter doc when available
+  - Store link preview info during build
 
 ## [4.8.0]